Support for reading config.json from JAR file
With JAR, there arises the necessity to execute the JAR from any
directory. So the config.json file must be bundled in JAR and the
application must read this file.

This commit adds support for JAR config.json reading (executed from any
directory, of course)

Also, minor improvement of README

Signed-off-by: Daniel Santos <hello@brighterdan.com>