able to read vars endpoint, apikey, secretapikey
the variables endpoint, apikey and secretapikey were only being read
from the config.json. When running as jar, it compromises security
because the API secrets able exposed inside the config jar file.

To mitigate this, all of this variables can now be set inside the
.java file which in turn produces a compiled .class file. Since this is
compiled, it is harder to read such API variables, making it more safe
to produce JAR files.

Also, now it can be optional to have an config.json file

Signed-off-by: Daniel Santos <hello@brighterdan.com>